requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.1
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.9
PySocks==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webdriver-manager==3.8.6
websockets==11.0.3
wsproto==1.2.0
